Update InfoAddictCivRelations.lua
The function showVisibleCivIcons contains the line that was causing the lua.log error Iris was receiving. My edit comments out the contents of the BuildView function, which is the only place in which showVisibleCivIcons gets called. I believe this change should result in the InfoAddict civ relations screen showing up as blank when visited, without affecting the remainder of InfoAddict's functionality