Update DMS_Ethiopia_Functions.lua
Streamlining code to take up less space. These changes are mainly intended to improve readability for humans and are not expected to change runtime significantly. Mostly (but not entirely) they involve tightening up the presentation of the Seed of Jacob promotions; I'll probably do another pass later to clean up more of the code around the dummy buildings